backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
requests>=2.31.0
httpx>=0.25.0
playwright>=1.40.0
python-multipart>=0.0.6
aiohttp>=3.9.0
python-dotenv>=1.0.0
